Backend/feature/feed api

See merge request s09-final/S09P31E203!80

+++ backend/timing/src/main/java/com/kkukku/timing/apis/member/services/MemberService.java
@@ -64,7 +64,9 @@ public class MemberService {
         MemberEntity member = getMemberById(memberId);
 
         String profileImgUrl = member.getProfileImageUrl();
-        s3Service.deleteFile(profileImgUrl);
+        if (!profileImgUrl.startsWith("http")) {
+            s3Service.deleteFile(profileImgUrl);
+        }
 
         member.delete();
